Extracted mp token fetching to separate class
